# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_bigops_platform_internal_model.User:
    properties:
      avatar:
        type: string
      created_at:
        example: "2024-01-01 00:00:00"
        type: string
      email:
        type: string
      id:
        type: integer
      phone:
        type: string
      real_name:
        type: string
      status:
        type: integer
      updated_at:
        example: "2024-01-01 00:00:00"
        type: string
      username:
        type: string
    type: object
  github_com_bigops_platform_internal_pkg_response.PageData:
    properties:
      list: {}
      page:
        description: 当前页码
        type: integer
      size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  github_com_bigops_platform_internal_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_handler.UpdateUserStatusRequest:
    properties:
      status:
        description: 0:禁用 1:启用
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: BigOps 运维平台 API 文档
  title: BigOps API
  version: "1.0"
paths:
...
  /users:
    get:
      description: 分页获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/github_com_bigops_platform_internal_model.User'
                        type: array
                    type: object
              type: object
        "500":
          description: 查询失败
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - 用户管理
  /users/{id}/delete:
    post:
      description: 软删除用户，不允许删除管理员（ID=1）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "400":
          description: 不允许删除管理员
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
...
  /users/{id}/status:
    post:
      consumes:
      - application/json
      description: 更新用户状态，0=禁用 1=启用，不允许禁用管理员（ID=1）
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateUserStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 启用/禁用成功
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "400":
          description: 不允许禁用管理员
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/github_com_bigops_platform_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 启用/禁用用户
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    description: 'Bearer token 认证，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"